Rename audio library to AudioSwitch
