+++ app/src/main/java/com/twilio/video/app/ui/room/AudioViewState.kt
@@ -1,6 +1,6 @@
 package com.twilio.video.app.ui.room
 
-import com.twilio.audio_router.AudioDevice
+import com.twilio.audioswitch.AudioDevice
 
 data class AudioViewState(
     val selectedDevice: AudioDevice? = null,
