+++ audioswitch/gradle.properties
