+++ audioswitch/proguard-rules.pro
