+++ audioswitch/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.twilio.audio_router" >
+    package="com.twilio.audioswitch" >
 
     <uses-permission android:name="android.permission.BLUETOOTH" />
