+++ audioswitch/src/main/java/com/twilio/audioswitch/AudioDevice.java
@@ -1,4 +1,4 @@
-package com.twilio.audio_router;
+package com.twilio.audioswitch;
 
 import androidx.annotation.NonNull;
